fix(scrapling_fotocasa): metros must be int (dbt staging casts to INTEGER)
raw_data->>'metros' ~ '^200\.0$' breaks ::INTEGER cast in stg_fotocasa.sql.
Save as int from the regex group instead of float. Existing affected rows
fixed in DB via one-shot UPDATE jsonb_set.

dbt run after fix: INSERT 0 44 (44 new leads merged into dim_leads).

+++ scrapers/scrapling_fotocasa.py
@@ -232,12 +232,12 @@ class ScraplingFotocasa(ScraplingBaseScraper):
             except Exception:
                 pass
 
-        # Surface: <number> m² / m2
+        # Surface: <number> m² / m2 — INT (dbt stg_fotocasa casts to INTEGER)
         metros = None
         m2_m = re.search(r"(\d{2,4})\s*m[²2]", text)
         if m2_m:
             try:
-                metros = float(m2_m.group(1))
+                metros = int(m2_m.group(1))
             except Exception:
                 pass
